// Angular2_ABS/app/common/model/dataClass.model.ts
export class RssDataClass {
	di: any = 0;
}
//class bodyClass {
//	transfer: any = "01";
//	temperature: any = "대기";
//	humidity: any = "대기";
//	doorSensor: any = "대기";
//	doorLock: any = "대기";
//	fan: any = "대기";
//	subStation: any = "대기";
//	extFlag: any = "대기";
//	ups: any = "대기";
//	fanCompulsion: any = "대기";
//	fanAutomatic: any = "대기";
//	fanAutoTemperature: any = "대기";
//	transferAutomatic: any = "대기";
//	transferLossStd: any = "대기";
//	transferCompulsion: any = "대기";
//}
export class GaugeClass {
	name: string;
	mean: number;
	//min: number;
	subValue: any;
	endValue: any;
	trueValue: any;
	release: any;
	tick: any;
}